# adapt/docs/openapi.json
{
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "LoginBody": {
    "properties": {
     "password": {
      "maxLength": 256,
      "minLength": 1,
      "title": "Password",
      "type": "string"
     },
     "user_id": {
      "maxLength": 64,
      "minLength": 1,
      "title": "User Id",
      "type": "string"
     }
    },
    "required": [
     "user_id",
     "password"
    ],
    "title": "LoginBody",
    "type": "object"
   },
...
   "SessionOut": {
    "properties": {
     "auth_enabled": {
      "title": "Auth Enabled",
      "type": "boolean"
     },
     "csrf_token": {
      "title": "Csrf Token",
      "type": "string"
     },
     "user": {
      "$ref": "#/components/schemas/SessionUser"
     }
    },
    "required": [
     "auth_enabled",
     "user",
     "csrf_token"
    ],
    "title": "SessionOut",
    "type": "object"
   },
   "SessionUser": {
    "properties": {
     "display_name": {
      "title": "Display Name",
      "type": "string"
     },
     "role": {
      "title": "Role",
      "type": "string"
     },
     "user_id": {
      "title": "User Id",
      "type": "string"
     }
    },
    "required": [
     "user_id",
     "display_name",
     "role"
    ],
    "title": "SessionUser",
    "type": "object"
   },
...
   "ValidationError": {
    "properties": {
     "ctx": {
      "title": "Context",
      "type": "object"
     },
     "input": {
      "title": "Input"
     },
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   },
...
  }
 },
 "info": {
  "title": "ADAPT API",
  "version": "0.1.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/v1/auth/login": {
   "post": {
    "operationId": "login_api_v1_auth_login_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/LoginBody"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SessionOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Login",
    "tags": [
     "auth"
    ]
   }
  },
...
 }
}